+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to wasm-vips will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.0.8] - TBD
+## [v0.0.8] - 2024-03-17
 
 Uses libvips v8.15.2, compiled with Emscripten v3.1.56.